+++ src/Command/BuiltIn/Releases/ListCommand.php
@@ -74,6 +74,9 @@ class ListCommand extends AbstractCommand
             $output->writeln('');
 
             $hosts = $this->runtime->getEnvOption('hosts');
+            if (!is_array($hosts) && !$hosts instanceof \Countable) {
+                $hosts = [];
+            }
             if (count($hosts) == 0) {
                 $output->writeln('No hosts defined');
                 $output->writeln('');