+++ src/MageApplication.php
@@ -84,6 +84,9 @@ class MageApplication extends Application
 
                 $logger = new Logger('magephp');
                 $logger->pushHandler(new StreamHandler($logfile));
+
+            } elseif (array_key_exists('log_dir', $config['magephp']) && !is_dir($config['magephp']['log_dir'])) {
+                throw new RuntimeException(sprintf('The configured log_dir "%s" does not exists or is not a directory.', $config['magephp']['log_dir']));
             }
 
             $this->runtime->setConfiguration($config['magephp']);